requirements.txt
opencv-python>=3.2.0.6
jupyter>=1.0.0
matplotlib>=2.1.1
pandas>=0.22.0
numpy>=1.12.1
pillow>=5.0.0
scipy>=1.0.0
nltk>=3.2.2
tqdm>=4.19.4
scikit-learn>=0.19.1
scikit-image>=0.13.1
seaborn>=0.8.1
streamlit>=0.79.0
torch>=0.4.0
torchvision>=0.2.0
Cython
pycocotools
